articulation-generation/requirements.txt
pybullet
open3d
matplotlib
scipy
transformations
numpy-quaternion
ConfigArgParse
tqdm
munch
opencv-python
pyrealsense2